@chrismclarke 2022-12-08
Thanks for the review and creating the follow-up issues @jfmcquade , will merge now